requirements.txt
# Software Engineering Review Crew Dependencies
# Core CrewAI dependencies
crewai>=0.80.0
crewai-tools>=0.15.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Code analysis tools
gitpython>=3.1.40
ast-grep-py>=0.12.0
radon>=6.0.1
bandit>=1.7.5
flake8>=7.0.0
pylint>=3.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0